Move ServiceAtomBehaviour related classes to its own extension package
