+++ webofneeds/won-bot/src/main/java/won/bot/framework/extensions/serviceatom/ServiceAtomContext.java
@@ -1,4 +1,4 @@
-package won.bot.framework.bot.context;
+package won.bot.framework.extensions.serviceatom;
 
 import java.net.URI;
