+++ webofneeds/won-bot/src/main/java/won/bot/framework/extensions/serviceatom/ServiceAtomExtension.java
@@ -1,0 +1,10 @@
+package won.bot.framework.extensions.serviceatom;
+
+@FunctionalInterface
+public interface ServiceAtomExtension {
+    /**
+     * The Behaviour defining this extension. For an example, see
+     * ServiceAtomBehaviour
+     */
+    ServiceAtomBehaviour getServiceAtomBehaviour();
+}
